Fix Fyne threading error by using async Do() instead of DoAndWait()
The setContent function was calling fyne.DoAndWait() from the main goroutine,
which created a deadlock. Changed to use fyne.Do() (asynchronous) to properly
marshal UI updates without blocking.

This resolves the error:
  'fyne.Do[AndWait] called from main goroutine'

The async approach is correct here since we don't need to wait for the
content update to complete before continuing.